Move conversions to mitiq.interface.
