+++ mitiq/interface/__init__.py
@@ -13,6 +13,16 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+from mitiq.interface.conversions import (
+    accept_any_qprogram_as_input,
+    atomic_converter,
+    atomic_one_to_many_converter,
+    convert_from_mitiq,
+    convert_to_mitiq,
+    noise_scaling_converter,
+    CircuitConversionError,
+    UnsupportedCircuitError,
+)
 from mitiq.interface import (
     mitiq_braket,
     mitiq_cirq,
